auth system added

+++ wcd_mainapp/views.py
@@ -3,7 +3,7 @@ from django.shortcuts import get_object_or_404, redirect, render
 from django.contrib import messages
 from wcd_mainapp.models import Tasks
 from wcd_mainapp.forms import *
-# from loguru import logger
+from django.contrib.auth.decorators import login_required
 from wcd_mainapp.tasks import periodic_task_scheduler, task_scheduler
 
 # Create your views here.
@@ -12,20 +12,21 @@ from wcd_mainapp.tasks import periodic_task_scheduler, task_scheduler
 def home(request):
     return render(request, 'wcd_mainapp/home.html')
 
+@login_required
 def all_tasks(request):
     context = {
-        'tasks':Tasks.objects.all()
+        'tasks':Tasks.objects.filter(user = request.user)
     }
     return render(request, 'wcd_mainapp/tasks.html', context)
 
+@login_required
 def add_tasks(request):
     if request.method == 'POST':
         add_form = TasksCreateForm(request.POST or None)
         if add_form.is_valid():
+            add_form.instance.user = request.user
             form_saved = add_form.save()
-            print(add_form.cleaned_data)
             data = add_form.cleaned_data
-            # task_scheduler.delay(form_saved.pk, data['web_url'], 1, 1.0, "full")    # for celery task (demo data for now)
             task_scheduler.delay(request.user.email, form_saved.pk, data['web_url'], data['detection_type'], data['threshold'], data['partOf'])    # for celery task
             messages.success(request, f"Your Task details has been saved!")
         return redirect('all_tasks')
@@ -39,31 +40,37 @@ def add_tasks(request):
     return render(request, 'wcd_mainapp/add_tasks.html', context)
     
 
+@login_required
 def update_tasks(request, id):
     if request.method == 'POST':
         instance = get_object_or_404(Tasks, id=id)
+
+        if not request.user == instance.user:
+            return HttpResponse(status=403)
+
         update_form = TasksUpdateForm(request.POST, instance=instance)
         if update_form.is_valid():
             update_form.save()
             periodic_task_scheduler.delay(request.user.email) # for celery beat (scheduled task)
+            messages.success(request, f"Your Task details has been updated!")
             return HttpResponse(status=200)
-            # messages.success(request, f"Your Task details has been updated!")
+            
         else:
+            messages.error(request, f"Update failed!")
             return HttpResponse(status=400)
-            # messages.error(request, f"Update failed!")
-
+            
     return HttpResponse(status=405)
     
 
+@login_required
 def delete_tasks(request, id):
     if request.method == "DELETE":
-        # if request.user:
         task = get_object_or_404(Tasks, id=id)
+        if not request.user == task.user:
+            return HttpResponse(status=403)
         task.delete()
-        # messages.success(request, f"Deletion Successful!")
+        messages.success(request, f"Deletion Successful!")
         return HttpResponse(status=200)
-        # else:
-            # return HttpResponse(status=403)
     else:
         return HttpResponse(status=405)
